requirements.txt
fastapi==0.95.0
uvicorn
fastapi-router-controller
pyclear
kubernetes==25.3.0
zope.interface
zope.component